requirements.txt
requests>=2.28
python-dotenv>=1.0
# Optional - if you intend to use OpenAI for final synthesis:
openai>=0.27.0
PyPDF2
streamlit
tavily
python-dotenv
matplotlib
openai
